new battle log system/ gen6 battletype effects/ other random changes(onboardpokemon.types...)

+++ trunk/src/PBO.Game.Host/Triggers/Lv7DChanging.cs
@@ -50,7 +50,7 @@ namespace PokemonBattleOnline.Game.Host.Triggers
         if (showFail)
         {
           pm.RaiseAbility();
-          pm.AddReportPm("7DLock", stat);
+          pm.AddReportPm("7DLock", (int)stat);
         }
         change = 0;
       }
